refactor(releve): index de compteur en Integer (contrat electricore = kWh entiers) (#133)
* refactor(releve): index de compteur en Integer (#132)

electricore-client 0.2.0 déclare index_*_kwh: int | None (ADR-0034 côté
electricore : floor Wh→kWh au boundary d'ingestion, la résolution sub-kWh
n'existe plus). Les 7 champs index_* de souscription.releve passent de
Float à Integer ; les énergies/montants restent Float, seuls les index sont
concernés.

* refactor(periode): defaut index absent en 0 entier (#132)

_releve_vals_depuis_objet projette les 7 index_*_kwh vers des champs
Integer désormais (souscription.releve) : le défaut `or 0.0` devient
`or 0`, cohérent avec le type de champ.

* refactor(reports): simplifie l'affichage des index en sortie directe (#132)

Le format '%g' du justificatif PDF masquait déjà les décimales d'un index
Float ; devenu redondant maintenant que index_* est Integer (t-out direct,
comme le fait déjà le portail).

* refactor(releve): migration pre-migrate 19.0.1.7.0 pour la colonne integer (#132)

Odoo ne convertit pas double precision -> integer sur un changement de type
de champ : sans migration, -u renomme les 7 colonnes index_* (perte
silencieuse). Ajoute migrations/19.0.1.7.0/pre-migrate.py (ALTER ... USING
round(...)::integer, idempotent) et bump la version du manifeste en
19.0.1.7.0 pour déclencher le script.

* test(portal): verrouille le rendu entier des index au portail (#132)

Quickfix de revue : symétrie avec le test PDF — assertNotIn('.0') pour
qu'un rendu flottant résiduel fasse échouer le test portail aussi.


---------

+++ models/core/souscription_releve.py
@@ -42,13 +42,17 @@ class SouscriptionReleve(models.Model):
     # Index par cadran réseau (même axe mesuré que energie_*) — colonnes
     # pertinentes selon config_cadrans : 4_cadrans → HPH/HPB/HCH/HCB ;
     # hp_hc → HP/HC ; base → Base.
-    index_hph = fields.Float(string='Index HPH', help='Heures Pleines saison Haute')
-    index_hpb = fields.Float(string='Index HPB', help='Heures Pleines saison Basse')
-    index_hch = fields.Float(string='Index HCH', help='Heures Creuses saison Haute')
-    index_hcb = fields.Float(string='Index HCB', help='Heures Creuses saison Basse')
-    index_hp = fields.Float(string='Index HP')
-    index_hc = fields.Float(string='Index HC')
-    index_base = fields.Float(string='Index Base')
+    # Integer (#132) : le contrat electricore (index_*_kwh) normalise l'index en
+    # kWh entier par floor au boundary d'ingestion (ADR-0034 côté electricore) ;
+    # contrairement aux énergies (`energie_*_kwh`) et montants (`*_eur`), qui
+    # restent Float, la résolution sub-kWh n'existe plus pour un index.
+    index_hph = fields.Integer(string='Index HPH', help='Heures Pleines saison Haute')
+    index_hpb = fields.Integer(string='Index HPB', help='Heures Pleines saison Basse')
+    index_hch = fields.Integer(string='Index HCH', help='Heures Creuses saison Haute')
+    index_hcb = fields.Integer(string='Index HCB', help='Heures Creuses saison Basse')
+    index_hp = fields.Integer(string='Index HP')
+    index_hc = fields.Integer(string='Index HC')
+    index_base = fields.Integer(string='Index Base')
 
     # Provenance du justificatif côté electricore (#76, ADR 0020 §6) :
     # `releve_externe_id` (← `releve_id` du contrat v3) identifie le relevé
